# Conflicts:
#	presentation/ui/src/main/kotlin/com/berlin/aflami/navigation/NavGraph.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/navigation/routes/CastScreenRoute.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/navigation/routes/SearchByCountryScreenRoute.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/home/ContinueWatchingScreen.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/home/HomeScreen.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/search/actor/SearchByActorNameScreen.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/search/country/SearchByCountryScreen.kt
#	presentation/ui/src/main/kotlin/com/berlin/aflami/screens/search/search/SearchScreen.kt

+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/searchcountry/SearchByCountryViewModel.kt
@@ -89,11 +89,12 @@ class SearchByCountryViewModel(
         _state.update { it.copy(dropDownExpanded = false) }
     }
 
-    override fun onMovieClicked(movieId: Int) {
+    override fun onMovieClicked(movieId: Long) {
         sendNewEffect(SearchByCountryEffect.NavigatedToMovieDetailsScreen(movieId,"MOVIE"))
     }
 
     override fun onBackClicked() {
         sendNewEffect(SearchByCountryEffect.NavigatedBack)
     }
+
 }
